feat(nvim, wezterm): Update Catppuccin theme settings
This commit updates the Catppuccin theme settings in both Neovim and
Wezterm configurations. In Neovim, the theme flavour is set to auto and
the background is configured for both light and dark modes. The
transparent_background and term_colors options are turned off. The
dap and dap_ui integrations are enabled. In Wezterm, the theme names
are updated to match the new settings.