feat: Update Neovim and zsh configurations
This commit includes changes to the Neovim and zsh configuration files.
In Neovim, the 'default_integrations' option has been enabled. In the
zsh configuration, several changes have been made including the
removal of the 'oh-my-zsh' dependency, the addition of lazy loading
for NVM, and the inclusion of zsh-syntax-highlighting. The changes
are aimed at improving the overall user experience and performance.